refactor(feature-info): move code from service to composable
